+++ include/engine/graph/RenderAOV.h
@@ -25,6 +25,7 @@ namespace engine::graph {
     virtual RenderResourceType resourceType() const = 0;
     virtual RenderResourceFormat resourceFormat() const = 0;
     virtual bool usesRasterTargetSampling() const = 0;
+    virtual bool usesRaytracerPassState() const = 0;
     virtual bool supportsExecutor(RenderExecutorKind executor) const = 0;
     virtual void configureProducerPass(RenderPassNode& pass) const;